Before first test of function
I have debugging steps in place for the insertion of the first metric. I am about to test it for the first time via my local Postman. I also realized the numUsers column was obsolete as it would be better to just do a count(distinct()) with a where to get the number of Interested users.